% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbcdtutils-package.R
\docType{package}
\name{fbcdtutils-package}
\alias{fbcdtutils}
\alias{fbcdtutils-package}
\title{fbcdtutils: data.table Utilities for FBC Studies}
\description{
Provides functions and methods that rely on
    data.table to complement the fbcutils package.
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikmart@mikkomarttila.com}

}
\keyword{internal}
